requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
attrs==25.4.0
bcrypt==4.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
frozenlist==1.8.0
geoip2==5.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jwt==1.4.0
limits==5.6.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
maxminddb==2.8.2
mdurl==0.1.2
multidict==6.7.0
Naked==0.1.32
packaging==25.0
passlib==1.7.4
phonenumbers==9.0.14
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
status==0.2.5
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
